requirements.txt
python-dotenv
psycopg2-binary
pgvector
llama-index
openai
sqlalchemy
